Move files out of core
