+++ backend/infrahub/pools/default_allocator.py
@@ -3,9 +3,9 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Any
 
 from infrahub.core import registry
-from infrahub.core.node.pool.base import PoolAllocator
 from infrahub.core.protocols import CoreNumberPool
 from infrahub.exceptions import NodeNotFoundError, PoolExhaustedError, ValidationError
+from infrahub.pools.allocator import PoolAllocator
 
 if TYPE_CHECKING:
     from infrahub.core.attribute import BaseAttribute
